+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - [2026-05-16 22:29] home/godmode/config/prefill.json
 - [2026-05-16 22:33] home/godmode/config/env.sh
 - [2026-05-16 22:36] data/data/com.termux/files/home/godmode/lib/encoder.py
+- [2026-05-16 22:37] home/godmode/lib/parsers.py